refactor: 패키지 변경 및 코드 컨벤션 적용
