+++ src/main/java/chess/domain/board/BoardFactory.java
@@ -1,13 +1,13 @@
 package chess.domain.board;
 
-import chess.domain.pieces.Bishop;
-import chess.domain.pieces.piece.Color;
-import chess.domain.pieces.King;
-import chess.domain.pieces.Knight;
-import chess.domain.pieces.Pawn;
-import chess.domain.pieces.piece.Piece;
-import chess.domain.pieces.Queen;
-import chess.domain.pieces.Rook;
+import chess.domain.piece.Bishop;
+import chess.domain.piece.Color;
+import chess.domain.piece.King;
+import chess.domain.piece.Knight;
+import chess.domain.piece.Pawn;
+import chess.domain.piece.Piece;
+import chess.domain.piece.Queen;
+import chess.domain.piece.Rook;
 import chess.domain.square.File;
 import chess.domain.square.Rank;
 import chess.domain.square.Square;
