+++ src/main/java/chess/view/PieceView.java
@@ -1,7 +1,7 @@
 package chess.view;
 
-import chess.domain.pieces.piece.Color;
-import chess.domain.pieces.piece.Type;
+import chess.domain.piece.Color;
+import chess.domain.piece.Type;
 import java.util.Arrays;
 
 public enum PieceView {
