+++ src/test/java/chess/domain/piece/QueenTest.java
@@ -1,10 +1,8 @@
-package chess.domain.pieces;
+package chess.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import chess.domain.Movement;
-import chess.domain.pieces.piece.Color;
-import chess.domain.pieces.piece.Piece;
 import chess.domain.square.Square;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
