+++ src/test/java/chess/domain/piece/RookTest.java
@@ -1,29 +1,27 @@
-package chess.domain.pieces;
+package chess.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import chess.domain.Movement;
-import chess.domain.pieces.piece.Color;
-import chess.domain.pieces.piece.Piece;
 import chess.domain.square.Square;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-@DisplayName("킹")
-class KingTest {
+@DisplayName("룩")
+class RookTest {
 
     @DisplayName("움직일 수 있다")
     @Test
     void canMove() {
         //given
-        Square source = Square.from("e1");
-        Square destination = Square.from("e2");
+        Square source = Square.from("a1");
+        Square destination = Square.from("a3");
 
-        Piece king = new King(Color.WHITE);
+        Piece rook = new Rook(Color.WHITE);
         Movement movement = new Movement(source, destination);
 
         //when
-        boolean canMove = king.canMove(movement, null);
+        boolean canMove = rook.canMove(movement, null);
 
         //then
         assertThat(canMove).isTrue();
@@ -33,14 +31,14 @@ class KingTest {
     @Test
     void canNotMove() {
         //given
-        Square source = Square.from("e1");
-        Square destination = Square.from("e3");
+        Square source = Square.from("a1");
+        Square destination = Square.from("c3");
 
-        Piece king = new King(Color.WHITE);
+        Piece rook = new Rook(Color.WHITE);
         Movement movement = new Movement(source, destination);
 
         //when
-        boolean canMove = king.canMove(movement, null);
+        boolean canMove = rook.canMove(movement, null);
 
         //then
         assertThat(canMove).isFalse();
