+++ src/test/java/chess/view/PieceViewTest.java
@@ -3,8 +3,8 @@ package chess.view;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import chess.domain.pieces.piece.Color;
-import chess.domain.pieces.piece.Type;
+import chess.domain.piece.Color;
+import chess.domain.piece.Type;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
